refactor endpoints/search package
